07.25 SQL 및 ERD 수정

+++ src/main/webapp/WEB-INF/doc/공지사항/notice.sql
@@ -13,7 +13,7 @@ CREATE TABLE notice(
     nimage VARCHAR2(100),
     nimagesaved VARCHAR2(100),
     nsize LONG,
-    adminno NUMBER(10),
+    adminno INTEGER(10),
   FOREIGN KEY (adminno) REFERENCES admin (adminno)
 );
